chore(step-functions): split the asl file into multiple files (#437)
